requirements.txt
anthropic>=0.83.0
click>=8.1.0
flask>=3.0.0
httpx>=0.27.0
playwright>=1.58.0
rich>=13.0.0
trafilatura>=2.0.0
fpdf2
Pillow
